busstopp update

+++ de.bund.bfr.busstop/src/de/bund/bfr/busstopp/resources/ItemResource.java
@@ -11,6 +11,7 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.Response.ResponseBuilder;
 
 import de.bund.bfr.busstopp.Constants;
@@ -26,11 +27,14 @@ public class ItemResource {
 	UriInfo uriInfo;
 	@Context
 	Request request;
+    @Context
+    SecurityContext securityContext;
 	Long id;
 
-	public ItemResource(UriInfo uriInfo, Request request, Long id) {
+	public ItemResource(UriInfo uriInfo, Request request, SecurityContext securityContext, Long id) {
 		this.uriInfo = uriInfo;
 		this.request = request;
+		this.securityContext = securityContext;
 		this.id = id;
 	}
 
@@ -61,21 +65,27 @@ public class ItemResource {
 		ResponseX response = new ResponseX();
 		response.setId(id);
 		response.setAction("DELETE");
-		ItemLoader c = Dao.instance.getModel().get(id);
-		if (c != null) {
-			try {
-				c.delete();
-			} catch (IOException e) {
-				e.printStackTrace();
-				response.setSuccess(false);
-				response.setError(e.getMessage());
+		if (securityContext.isUserInRole("x2bfr")) {
+			ItemLoader c = Dao.instance.getModel().get(id);
+			if (c != null) {
+				try {
+					c.delete();
+				} catch (IOException e) {
+					e.printStackTrace();
+					response.setSuccess(false);
+					response.setError(e.getMessage());
+				}
+				c = Dao.instance.getModel().remove(id);
+				response.setSuccess(true);
 			}
-			c = Dao.instance.getModel().remove(id);
-			response.setSuccess(true);
+			else  {
+				response.setSuccess(false);
+				response.setError("ID not found");
+			}
 		}
-		else  {
+		else {
 			response.setSuccess(false);
-			response.setError("ID not found");
+			response.setError("No permission to access this feature!");
 		}
 		return response;
 	}
@@ -93,39 +103,59 @@ public class ItemResource {
 	@Path("file")
 	@Produces(MediaType.APPLICATION_OCTET_STREAM)
 	public Response getFile() {
-		ItemLoader c = Dao.instance.getModel().get(id);
-		if (c == null) return Response.noContent().build();
-		String filename = Constants.SERVER_UPLOAD_LOCATION_FOLDER + c.getXml().getId() + "/" + c.getXml().getIn().getFilename();
-		ResponseBuilder response = getDownloadResponse(filename);
-	    return response.build();
+		if (securityContext.isUserInRole("bfr2x")) {
+			ItemLoader c = Dao.instance.getModel().get(id);
+			if (c == null) return Response.noContent().build();
+			String filename = Constants.SERVER_UPLOAD_LOCATION_FOLDER + c.getXml().getId() + "/" + c.getXml().getIn().getFilename();
+			ResponseBuilder response = getDownloadResponse(filename);
+		    return response.build();
+		}
+		else {
+			return Response.noContent().build();
+		}
 	}
 	@GET
 	@Path("workflow")
 	@Produces(MediaType.APPLICATION_OCTET_STREAM)
 	public Response getWorkflow() {
-		ItemLoader c = Dao.instance.getModel().get(id);
-		if (c == null) return Response.noContent().build();
-		String filename = Constants.SERVER_UPLOAD_LOCATION_FOLDER + c.getXml().getId() + "/" + c.getXml().getOut().getWorkflow();
-		ResponseBuilder response = getDownloadResponse(filename);
-		return response.build();
+		if (securityContext.isUserInRole("bfr2x")) {
+			ItemLoader c = Dao.instance.getModel().get(id);
+			if (c == null) return Response.noContent().build();
+			String filename = Constants.SERVER_UPLOAD_LOCATION_FOLDER + c.getXml().getId() + "/" + c.getXml().getOut().getWorkflow();
+			ResponseBuilder response = getDownloadResponse(filename);
+			return response.build();
+		}
+		else {
+			return Response.noContent().build();
+		}
 	}
 	@GET
 	@Path("report")
 	@Produces(MediaType.APPLICATION_OCTET_STREAM)
 	public Response getReport() {
-		ItemLoader c = Dao.instance.getModel().get(id);
-		if (c == null) return Response.noContent().build();
-		String filename = Constants.SERVER_UPLOAD_LOCATION_FOLDER + c.getXml().getId() + "/" + c.getXml().getOut().getReport();
-		ResponseBuilder response = getDownloadResponse(filename);
-	    return response.build();
+		if (securityContext.isUserInRole("bfr2x")) {
+			ItemLoader c = Dao.instance.getModel().get(id);
+			if (c == null) return Response.noContent().build();
+			String filename = Constants.SERVER_UPLOAD_LOCATION_FOLDER + c.getXml().getId() + "/" + c.getXml().getOut().getReport();
+			ResponseBuilder response = getDownloadResponse(filename);
+		    return response.build();
+		}
+		else {
+			return Response.noContent().build();
+		}
 	}
 	@GET
 	@Path("comment")
 	@Produces(MediaType.TEXT_PLAIN)
 	public String getComment() {
-		ItemLoader c = Dao.instance.getModel().get(id);
-		if (c != null) {
-		    return c.getXml().getOut().getComment();
+		if (securityContext.isUserInRole("bfr2x")) {
+			ItemLoader c = Dao.instance.getModel().get(id);
+			if (c != null) {
+			    return c.getXml().getOut().getComment();
+			}
+			else {
+				return "";
+			}
 		}
 		else {
 			return "";
@@ -135,7 +165,12 @@ public class ItemResource {
 	@Path("zip")
 	@Produces(MediaType.TEXT_PLAIN)
 	public String getAllZipped() {
-		return "todo...";
+		if (securityContext.isUserInRole("bfr2x")) {
+			return "todo...";
+		}
+		else {
+			return "";
+		}
 	}
 	
 }
